blog backend Commit

+++ Backend/src/main/java/com/Ajinkya/Infotech/repository/OtpRepository.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Repository
-public interface OtpRepository extends JpaRepository<Otp,Integer> {
+public interface OtpRepository extends JpaRepository<Otp,Long> {
     Optional<Otp> findByEmail(String email);
 
     void deleteByEmail(String email);
